MLK-11920 video: fbdev: mxc ipuv3 fb: Use fb_set_var() to unblank a framebuffer
The resolution of an overlay framebuffer could be at most the same to the
one of the relevant background framebuffer.  However, the resolution of a
HDMI monitor could be changed at runtime.  Thus, when we unblank the
overlay framebuffer, we should check it's resolution and start position in
case it exceeds the boundary of the background framebuffer.  This patch
replaces mxcfb_set_par() with fb_set_var() to implement the unblank
operation so that the logic contains the ->fb_check_var() path.

Signed-off-by: Liu Ying <Ying.Liu@freescale.com>
Signed-off-by: Vipul Kumar <vipul_kumar@mentor.com>