MLK_12886-2 video: mxsfb: handle the assert gpio in driver to support deferred probe
The assert gpio comes from 'gpio_spi' module, so the framebuffer
depends on the 'gpio_spi' driver loading. And in the case that
the framebuffer driver is loaded earlier than the 'gpio_spi'
driver, the gpio asserting will fail. So handle this gpio in
the framebuffer driver and add deferred probed support.

Signed-off-by: Fancy Fang <chen.fang@nxp.com>
(Vipul: FIxed merge conflicts)
Signed-off-by: Vipul Kumar <vipul_kumar@mentor.com>