MLK-21222 rpmsg: imx: fix kernel dump when initialize the vring memory
The fix had been contained in commit "b7098fc4" MLK-16371-2 rpmsg: imx:
add the rpmsg for imx8qxp on imx_4.14.y branch.
But part of that patch is missed. Merge the solution back to imx_4.19.y
branch.
Use memset_io replace memset, because that the memory type of
rpvq->addr is strong order type. There would be kernel dump when memset
is used on imx8qxp.

Signed-off-by: Richard Zhu <hongxing.zhu@nxp.com>