MLK-18795-1: Revert "MLK-17089-3: ASoC: wm8960: support suspend & resume for imx8"
Clock management is already implemented in wm8960_set_bias_level_out3,
there is no need to do it in runtime suspend/resume. Another aspect is
that the component_driver.idle_bias_on = 1, because of this codec will
be in runtime_resume all the time and runtime_suspend will never be called,
with current code the implication of this is that the clock will be
enabled all the time even codec not in use.

This reverts commit d845971fdd5531a1f3c37bbc45560aedd3da0773.

Signed-off-by: Viorel Suman <viorel.suman@nxp.com>