MLK-22097: VPU Decoder: speed up firmware boot after suspend
If there is no any decode instance, after suspend,
firmware will boot again,
and before boot, it will cleanup memory,
the memory is large for mu, so it cost a lot of time.
If cpu cleanup this memory, and firmware skip this step,
the boot time can be greatly reduced.

Signed-off-by: ming_qian <ming.qian@nxp.com>
Reviewed-by: Shijie Qin <shijie.qin@nxp.com>